requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-telegram-bot
python-dotenv
psycopg2-binary
asyncpg
alembic
requests